Update kbScraper.py
Program functions as desired. I've tested it by adding and subtracting urls/names to and from the sheetNames and listOfURLs lists.
 
A minor snag is that the program takes a long time to run, about a minute from clicking the run button to the creation of the spreadsheet and printing "end" to the console. Is this because it takes time to query the website for the data? Is the target website slow? Is my code slow? Is my computer slow? I don't know. 

Future features to add:
-A column showing whether that product is on sale.
-Miscellaneous sheet styling. Cell colors, borders, etc.