Update kbScraper.py
Program functions as desired. I tested it by adding and removing URLs, the program dynamically creates the sheets and populates the desired data. 

Small hiccup, the program takes a long time to run, around a minute or so from clicking the run button to the spreadsheet creation and printing "end" to the console. I don't know why. Does it take that much time to query the website pages and extract the data? Is my computer slow? Is my code slow?

Features to be added later:
-an additional column next to the price to signify whether the product is currently on sale or out of stock.
-Miscellaneous styling. Cell colors, borders, etc.